Refactor general accessibility tests with DashboardSteps class
The GeneralSteps class was getting too large.

Issue #518 Implement accessibility checking in CI
