+++ psm-app/integration-tests/src/test/java/gov/medicaid/features/general/steps/DashboardStepDefinitions.java
@@ -8,48 +8,50 @@ import net.thucydides.core.annotations.Steps;
 public class DashboardStepDefinitions {
     @Steps
     GeneralSteps generalSteps;
+    @Steps
+    DashboardSteps dashboardSteps;
 
     @Given("^I am on the Dashboard Draft page$")
     public void i_am_on_the_dashboard_draft_page() {
         generalSteps.loginAsProvider();
-        generalSteps.openDashboardDraftPage();
+        dashboardSteps.openDashboardDraftPage();
     }
 
     @When("^I open the filter panel on the Dashboard Draft page")
     public void i_open_the_filter_panel_on_the_dashboard_draft_page() {
-        generalSteps.clickDashboardDraftPageFilterButton();
+        dashboardSteps.clickDashboardDraftPageFilterButton();
     }
 
     @Given("^I am on the Dashboard Pending page$")
     public void i_am_on_the_dashboard_pending_page() {
         generalSteps.loginAsProvider();
-        generalSteps.openDashboardPendingPage();
+        dashboardSteps.openDashboardPendingPage();
     }
 
     @When("^I open the filter panel on the Dashboard Pending page")
     public void i_open_the_filter_panel_on_the_dashboard_pending_page() {
-        generalSteps.clickDashboardPendingPageFilterButton();
+        dashboardSteps.clickDashboardPendingPageFilterButton();
     }
 
     @Given("^I am on the Dashboard Approved page$")
     public void i_am_on_the_dashboard_approved_page() {
         generalSteps.loginAsProvider();
-        generalSteps.openDashboardApprovedPage();
+        dashboardSteps.openDashboardApprovedPage();
     }
 
     @When("^I open the filter panel on the Dashboard Approved page")
     public void i_open_the_filter_panel_on_the_dashboard_approved_page() {
-        generalSteps.clickDashboardApprovedPageFilterButton();
+        dashboardSteps.clickDashboardApprovedPageFilterButton();
     }
 
     @Given("^I am on the Dashboard Denied page$")
     public void i_am_on_the_dashboard_denied_page() {
         generalSteps.loginAsProvider();
-        generalSteps.openDashboardDeniedPage();
+        dashboardSteps.openDashboardDeniedPage();
     }
 
     @When("^I open the filter panel on the Dashboard Denied page")
     public void i_open_the_filter_panel_on_the_dashboard_denied_page() {
-        generalSteps.clickDashboardDeniedPageFilterButton();
+        dashboardSteps.clickDashboardDeniedPageFilterButton();
     }
 }
